chore: planner cleanup and refactor
